+++ sdk/CHANGELOG.md
@@ -4,6 +4,8 @@ This changelog summarizes major changes between Graal SDK versions. The main foc
 
 ## Version 1.0 RC9
 * Added a [Context.Builder.logHandler](http://www.graalvm.org/sdk/javadoc/org/graalvm/polyglot/Context.Builder.html#logHandler-java.io.OutputStream-) and [Engine.Builder.logHandler](http://www.graalvm.org/sdk/javadoc/org/graalvm/polyglot/Engine.Builder.html#logHandler-java.io.OutputStream-) methods to install a logging handler writing into a given `OutputStream`
+* `Value.asValue(Object)` now also works if no currently entered context is available.
+* Primitives, host and `Proxy` values can now be shared between multiple context and engine instances. They no longer throw an `IllegalArgumentException` when shared. Primitive types are `Boolean`, `Byte`, `Short`, `Integer`, `Long`, `Float`, `Double`, `Character` and `String` of the `java.lang` package. Non primitive values originating from guest languages are not sharable.
 
 ## Version 1.0 RC8
 * Added `MessageTransport` and `MessageEndpoint` to virtualize transport of messages to a peer URI.